requirements.txt
questionary
flask
flask-cors
uvicorn
pycryptodome
requests
python-dotenv
python-multipart
texttable